pyproject.toml
[project]
name = "mermaid-parser-py"
version = "0.0.1"
description = "Parse Mermaid into manipulable Python objects"
authors = [
    {name = "Boqi Chen",email = "boqi.chen@mail.mcgill.ca"}
]
repository = "https://github.com/20001LastOrder/mermaid-parser-py"
license = {text = "MIT"}
readme = "README.md"
requires-python = "<4.0,>=3.11"
dependencies = [
    "mermaid-py (>=0.8.0,<0.9.0)",
    "pythonmonkey (>=1.2.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "networkx (>=3.5,<4.0)",
    "loguru (>=0.7.3,<0.8.0)"
]

[tool.poetry]
packages = [
    { include = "mermaid_parser" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

